Define chips for letter lookups...
The 'i' pins combined denote a binary number
between 0 and 63.

The 'o' pins combined denote a binary number
between 0 and 7.

Let's take an example from the 'e_lookup' chip:

| 0 | 0 | 0 | 1 | 0 | 1 | 0 | 1 | 0 |

This row maps the input number 5 to the output
number 2.

This is because the phrase "five e's" contains two
'e' characters.

Note that the 's_lookup' chip is special, because
of the trailing 's'. It adds an extra 's' to every
number except one, which would not result in a
trailing 's'.