Check-in the compiled pangram chip...
This is the compilation of the top-level pangram
chip into DIMACS format. This took ~45 mins to
compute and required ~2.5GB of memory.

This file is effectively a reduction of the
self-enumerating pangram problem into a the
boolean satisfiability problem.

The final stage is to provide a wrapper that does
encoding - solving - decoding of specific seed
frequencies provided by the user.