Define the e_number chip…
This chip returns the number of ‘e’ letters that
appear in the percentages.

All of these are valid:

1.23 -> one point two three
1.20 -> one point two
1.00 -> one

Because of this, it conditionally chops off the
second and third significant figures depending on
trailing zeroes.

This commit also includes some conditional logic
to account for the word “point” in the phrase, but
this isn’t tested yet because ‘e’ does not appear
in the word “point”.