Make it so that a_equal is true…
This is useful for debugging purposes. Given the
other letters in the patuzzogram chip are copies
of the ‘a’ schematic - this is the closest we can
get to verifying that the chip works as expected.

There are currently no known solutions for this
word problem. It would make sense to replace the
quoted values with a real solution when one is
found.